Style-Improvements: Added mew cover. BookList component. Removed library page.
