+++ src/components/header.js
@@ -1,13 +1,20 @@
 import { Link } from "gatsby"
 import PropTypes from "prop-types"
 import React from "react"
+import Search from "../images/Icons/search.svg";
+import Settings from "../images/Icons/settings.svg";
 import "./header.scss";
 
 const Header = ({ siteTitle }) => (
   <header className="header">
-    <Link to="/library" className="link">
-      <img src="/icons/icon-72x72.png" />
-    </Link>
+    <div className="leftSideNavigation">
+      <span>English</span>
+      <img className="page-navigation-item" src="http://icons.iconarchive.com/icons/custom-icon-design/all-country-flag/256/United-Kingdom-flag-icon.png" alt="Italian Trulli" height="28"></img>
+    </div>
+    <div className="rightSideNavigation">
+      <img className="page-navigation-item" src={Search} alt="Italian Trulli"></img>
+      <img className="page-navigation-item" src={Settings} alt="Italian Trulli"></img>
+    </div>
   </header>
 )
